#119 updated list form test to include button row
